pyproject.toml
[project]
name = "oreilly-langchain-course"
version = "2.0.0"
description = "O'Reilly Live Training - Getting Started with LangChain 1.0+"
requires-python = ">=3.11,<3.14"

dependencies = [
    # LangChain 1.0+ Core
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-community>=0.4.0",
    "langgraph>=1.0.0",
    "langsmith>=0.4.0",

    # Search Tools
    "langchain-tavily>=0.1.0",

    # RAG Dependencies
    "langchain-chroma>=1.0.0",
    "langchain-text-splitters>=1.0.0",

    # Document Loaders
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",

    # Development
    "jupyter>=1.0.0",
    "python-dotenv>=1.0.0",
    "langgraph-cli[inmem]>=0.4.0",

    # Utilities
    "pydantic>=2.0.0",
]
